sync aws s3

+++ wp-content/plugins/wp-persistent-login/classes/wp-persistent-login.php
@@ -32,7 +32,7 @@ class WP_Persistent_Login {
 		add_filter( 'auth_cookie_expiration', array( $this, 'set_login_expiration' ), 10, 3 );
 
 		// increase the cookie time when a user revisits
-		add_action( 'auth_cookie_valid', array( $this, 'update_auth_cookie' ), 10, 2 );
+		add_action( 'set_current_user', array( $this, 'update_auth_cookie' ), 10, 0 );
 
 		// set user meta if the user want to be remembered
 		add_filter( 'secure_signon_cookie', array( $this, 'remember_me_meta' ), 20, 2 ); 
@@ -96,22 +96,28 @@ class WP_Persistent_Login {
 	 * @return bool
 	 */
 	public function remember_me_meta( $secure_cookie, $credentials ) { 
-				
-		$user = get_user_by('login', $credentials['user_login']);
 
-		if ( !empty( $user ) ) {			
-		
-			if( $credentials['remember'] === true ) {
-				update_user_meta( $user->ID, 'persistent_login_remember_me', 'true');
+		if( $credentials['user_login'] != null ) {
+
+			$user = get_user_by('login', $credentials['user_login']);
+
+			if ( !empty( $user ) ) {			
+			
+				if( $credentials['remember'] === true ) {
+					update_user_meta( $user->ID, 'persistent_login_remember_me', 'true');
+				}
+				
+				if( $credentials['remember'] === false ) {
+					delete_user_meta( $user->ID, 'persistent_login_remember_me', 'true');
+				}
+			
 			}
 			
-			if( $credentials['remember'] === false ) {
-				delete_user_meta( $user->ID, 'persistent_login_remember_me', 'true');
-			}
-		
+			return $secure_cookie; 
+
 		}
+				
 		
-		return $secure_cookie; 
 
 	} 
 
@@ -123,67 +129,53 @@ class WP_Persistent_Login {
 	 * Reset authentication cookie expiry to keep the user logged in
 	 *
 	 * @since  1.0.0
-	 * @param  array $cookieElements
-	 * @param  object $user
 	 * @return void
 	 */
-	public function update_auth_cookie( $cookieElements, $user ) {					
+	public function update_auth_cookie() {
+
+		$user = wp_get_current_user();
 				
-		if( $user ) :
+		if( !is_wp_error($user) && NULL !== $user ) :
 
 			// set users local cookie again - checks if they should be remembered
 			$remember_user_check = get_user_meta( $user->ID, 'persistent_login_remember_me', true );
+			$cookie = wp_parse_auth_cookie('', 'logged_in');
+			
+			// if there's no cookie, stop processing
+			if( !$cookie ) {
+				return;
+			}
+
+			$cookie_expiration = $cookie['expiration'];
+			$does_cookie_need_updating = $this->does_cookie_need_updating( $cookie_expiration );
 						
-			if( $remember_user_check === 'true' ) :
+			if( $remember_user_check === 'true' && $does_cookie_need_updating == true ) :
 				
 				// get the session verifier from the token
-					$sessionToken = $cookieElements['token'];
-						
-					if ( function_exists( 'hash' ) ) :
-						$verifier = hash( 'sha256', $sessionToken );
-					else :
-						$verifier = sha1( $sessionToken );
-					endif;						
-						
-				// update the login time, expires time
+					$session_token = $cookie['token'];
+					$verifier = $this->get_session_verifier_from_token( $session_token );		
+					
+				// get the current users sessions
 					$sessions = get_user_meta( $user->ID, 'session_tokens', true );
-					$sessions[$verifier]['login'] = time();
-					$sessions[$verifier]['expiration'] = time()+$this->expiration;
-					$sessions[$verifier]['ip'] = $_SERVER["REMOTE_ADDR"];
 						
-				// update the token with new data
-					$wp_session_token = WP_Session_Tokens::get_instance( $user->ID );
-					$wp_session_token->update( $sessionToken, $sessions[$verifier] );
-				
-				// apply filter for allowing duplicate sessions, default false
-					$currentOptions = get_option( 'persistent_login_options' );
-					$allowDuplicateSessions = $currentOptions['duplicateSessions'];
-						
-						
-				// remove any exact matches to this session
-					if( $allowDuplicateSessions === '0' ) :
+					if( $sessions != '' ) {
 
-						foreach( $sessions as $key => $session ) :
-							if( $key !== $verifier ) : 
-																												
-									// if we're on the same user agent and same IP, we're probably on the same device
-									// delete the duplicate session
-									if( 
-										($session['ip'] === $sessions[$verifier]['ip']) &&
-										($session['ua'] === $sessions[$verifier]['ua'])
-									) :
-										$updateSession = new WP_Persistent_Login_Manage_Sessions( $user->ID );
-										$updateSession->persistent_login_update_session( $key );
-									endif;
-																			
-							endif; // if key is different to identifier 
-						endforeach;
+						// update the login time, expires time if the user has sessions
+							$this->update_cookie_expiry($sessions, $session_token, $user->ID, $verifier);
 
-					endif;
+						// apply filter for allowing duplicate sessions, default false
+							$currentOptions = get_option( 'persistent_login_options' );
+							$allowDuplicateSessions = $currentOptions['duplicateSessions'];
+								
+						// remove any exact matches to this session
+							if( $allowDuplicateSessions === '0' ) :
+								$this->remove_duplicate_sessions($sessions, $verifier, $user->ID);
+							endif;
+
+					}				
 				
 				// if the user should be remembered, reset the cookie so the cookie time is reset
-					wp_set_auth_cookie( $user->ID, true, is_ssl(), $sessionToken );
-								
+					wp_set_auth_cookie( $user->ID, true, is_ssl(), $session_token );
 
 			endif; // end if remember me is set	
 
@@ -337,6 +329,109 @@ class WP_Persistent_Login {
 
 		$credentials['remember'] = true;
 		return $credentials;
+
+	}
+
+
+	/**
+	 * does_cookie_need_updating
+	 * 
+	 * Checks to see if the cookies was set less than a day ago
+	 * If it was, the cookie does not need to be updated.
+	 *
+	 * @param  array $cookieElements
+	 * @return bool
+	 * 
+	 * @since 2.0.11
+	 */
+	private function does_cookie_need_updating( $cookie_expiration = NULL ) {
+
+		if( !$cookie_expiration ) {
+			return true; // update the cookie if we don't know the expirtaion
+		}
+
+		$expiration_minus_one_day = time() + $this->expiration - 60; // DAY_IN_SECONDS
+		if( $cookie_expiration < $expiration_minus_one_day ) {
+			return true; // update the cookie if it was set more than a day ago
+		}
+
+		// otherwise, don't update the cookie
+		return false;
+
+	}
+
+	
+	/**
+	 * get_session_verifier_from_token
+	 *
+	 * @param  string $session_token
+	 * @return string
+	 */
+	private function get_session_verifier_from_token( $session_token ) {
+						
+		if ( function_exists( 'hash' ) ) :
+			$verifier = hash( 'sha256', $session_token );
+		else :
+			$verifier = sha1( $session_token );
+		endif;		
+
+		return $verifier;
+
+	}
+
+	
+	/**
+	 * update_cookie_expiry
+	 *
+	 * @param  array $sessions
+	 * @param  string $session_token
+	 * @param  int $user_id
+	 * @param  string $verifier
+	 * @return void
+	 */
+	private function update_cookie_expiry($sessions, $session_token, $user_id, $verifier) {
+
+		// update the login time, expires time
+		$sessions[$verifier]['login'] = time();
+		$sessions[$verifier]['expiration'] = time()+$this->expiration;
+		$sessions[$verifier]['ip'] = $_SERVER["REMOTE_ADDR"];
+			
+		// update the token with new data
+		$wp_session_token = WP_Session_Tokens::get_instance( $user_id );
+		$wp_session_token->update( $session_token, $sessions[$verifier] );
+
+	}
+
+	
+	/**
+	 * remove_duplicate_sessions
+	 *
+	 * @param  array $sessions
+	 * @param  string $verifier
+	 * @param  int $user_id
+	 * @return void
+	 */
+	private function remove_duplicate_sessions( $sessions, $verifier, $user_id ) {
+
+		foreach( $sessions as $key => $session ) :
+			if( $key !== $verifier ) : 
+																								
+					// if we're on the same user agent and same IP, we're probably on the same device
+					// delete the duplicate session
+					if( 
+						array_key_exists( $verifier, $sessions ) &&
+                        array_key_exists( 'ip', $sessions[$verifier] ) &&
+                        array_key_exists( 'ua', $sessions[$verifier] ) &&
+                        ($session['ip'] === $sessions[$verifier]['ip']) &&
+                        ($session['ua'] === $sessions[$verifier]['ua'])
+                    ) :
+						$updateSession = new WP_Persistent_Login_Manage_Sessions( $user_id );
+						$updateSession->persistent_login_update_session( $key );
+					endif;
+					
+															
+			endif; // if key is different to identifier 
+		endforeach;
 
 	}
